requirements.txt
discord>=1.0.1
psutil>=5.7.3
python-dotenv>=0.17
aiohttp>=3.7
